Refactoring menu fetching and csv parsing logic for reuse
