+++ website/src/components/menu.tsx
@@ -1,36 +1,28 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import { MenuItem as MenuItemComponent } from "@/components/menu-item";
-import Papa from "papaparse";
-import { fallbackMenu, fallbackMenuDate } from '@/lib/constants';
+import { MenuItemComponent } from "@/components/menu-item";
+import { fallbackMainMenu, fallbackMainMenuDate } from '@/lib/constants';
+import { fetchMenuData } from '@/data-fetcher/sheet-fetcher';
 
-interface CsvMenuDataRow {
-  category: 'FOOD' | 'DRINKS';
-  typeName: string;
-  typeDescription?: string;
-  itemName: string;
-  itemDescription?: string;
-  itemPrice?: number;
-  itemBulkPrice?: number;
-}
+
 
 export interface Menu {
   food: MenuCategory;
   drinks: MenuCategory;
 }
 
-interface MenuCategory {
+export interface MenuCategory {
   sections: MenuSection[];
 }
 
-interface MenuSection {
+export interface MenuSection {
   type: string;
   description?: string;
   items: MenuItem[];
 }
 
-interface MenuItem {
+export interface MenuItem {
   name: string;
   description?: string;
   price?: number;
@@ -39,11 +31,11 @@ interface MenuItem {
 
 const CACHE_KEY = "menuCache";
 const CACHE_EXPIRATION_KEY = "menuCacheExpiration";
-const CACHE_TTL = 1000 * 60 * 60 * 24; // 1 day expiration
+const MAIN_MENU_SHEET_LINK = "https://docs.google.com/spreadsheets/d/e/2PACX-1vQFe8zXVdVW7slOpUu8hsp32MnnEz1ZGRivhEWJjaBUIWxz5jRXd8qYjKrZ05KEQG0F-kT1YFlFiSaZ/pub?output=csv";
 
 export function Menu() {
   const [menuCategorySelection, setMenuCategorySelection] = useState<'FOOD' | 'DRINKS'>('FOOD');
-  const [menu, setMenu] = useState<Menu>(fallbackMenu);
+  const [menu, setMenu] = useState<Menu>(fallbackMainMenu);
 
   useEffect(() => {
     const urlParams = new URLSearchParams(window.location.search);
@@ -63,92 +55,7 @@ export function Menu() {
     }
 
     async function fetchData() {
-      try {
-        const dataUrl = "https://docs.google.com/spreadsheets/d/e/2PACX-1vQFe8zXVdVW7slOpUu8hsp32MnnEz1ZGRivhEWJjaBUIWxz5jRXd8qYjKrZ05KEQG0F-kT1YFlFiSaZ/pub?output=csv";
-        const dataResponse = await fetch(dataUrl, {
-          redirect: "follow",
-          cache: "no-store", // Prevent browser cache
-          headers: {
-            "Cache-Control": "no-cache, no-store, must-revalidate", // Ensure fresh request
-            "Pragma": "no-cache",  // HTTP/1.0 backward compatibility
-          }
-        });
-
-        if (!dataResponse.ok) throw new Error(`Failed to fetch CSV: ${dataResponse.statusText}`);
-
-        const csvText = await dataResponse.text();
-        Papa.parse<CsvMenuDataRow>(csvText, {
-          header: true,
-          skipEmptyLines: true,
-          dynamicTyping: false, // Prevent unintended type conversions
-          transform: (value, column) => {
-            if (column === "itemPrice" || column === "itemBulkPrice") {
-              return Number(value) || undefined;
-            }
-
-            return value;
-          },
-          error: (err: { message: string; }) => console.warn(`CSV Parsing Error: ${err.message}`),
-          complete: (result) => {
-            // Group by category and type first
-            const groupedData = result.data.reduce((acc, row) => {
-              // Create category if it doesn't exist
-              if (!acc[row.category]) {
-                acc[row.category] = {};
-              }
-
-              // Create type if it doesn't exist
-              if (!acc[row.category][row.typeName]) {
-                acc[row.category][row.typeName] = {
-                  items: [],
-                  description: row.typeDescription // Will take first occurrence
-                };
-              }
-
-              // Add item to the type
-              acc[row.category][row.typeName].items.push({
-                name: row.itemName,
-                description: row.itemDescription,
-                price: row.itemPrice,
-                bulkPrice: row.itemBulkPrice,
-              });
-
-              return acc;
-            }, {} as Record<string, Record<string, { items: MenuItem[], description?: string }>>);
-
-            // Transform into Menu structure with guaranteed food and drinks categories
-            const menu: Menu = {
-              food: {
-                sections: groupedData['FOOD'] ? Object.entries(groupedData['FOOD']).map(([typeName, data]) => ({
-                  type: typeName,
-                  description: data.description,
-                  items: data.items
-                })) : []
-              },
-              drinks: {
-                sections: groupedData['DRINKS'] ? Object.entries(groupedData['DRINKS']).map(([typeName, data]) => ({
-                  type: typeName,
-                  description: data.description,
-                  items: data.items
-                })) : []
-              }
-            };
-
-            // Store in localStorage with expiration
-            localStorage.setItem(CACHE_KEY, JSON.stringify(menu));
-            localStorage.setItem(CACHE_EXPIRATION_KEY, String(Date.now() + CACHE_TTL));
-
-            console.info('Fetched menu from source and updated cache.');
-
-            setMenu(menu);
-          },
-        });
-      } catch (e) {
-        console.info(`Fetching menu from source or cache failed (see error below), falling back to menu last updated on ${fallbackMenuDate}.`);
-        console.info(e);
-
-        setMenu(fallbackMenu);
-      }
+      fetchMenuData(MAIN_MENU_SHEET_LINK, setMenu, CACHE_KEY, CACHE_EXPIRATION_KEY, fallbackMainMenu, fallbackMainMenuDate);
     }
 
     fetchData();
